fix: enforce WorkItem terminal linkage at durability boundary
